swap data initialziation location from App to Main
